requirements.txt
# Python dependencies for KnightBot services
# Generated by Trae AI

# Core/Common
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
websockets>=11.0
httpx>=0.24.0

# LLM & Memory
langchain>=0.0.267
mem0ai
qdrant-client
openai

# Audio Processing (STT/TTS)
torch>=2.0.0
torchaudio>=2.0.0
soundfile
numpy
librosa
faster-whisper
scipy

# Parakeet STT (NeMo)
nemo-toolkit>=2.0.0

# Chatterbox TTS
transformers>=4.40.0
pydub>=0.25.1
python-multipart>=0.0.9

# LiveKit / Pipecat
livekit>=0.9.0
livekit-api>=0.4.0
livekit-agents>=0.5.0
pipecat-ai

# Image Processing
Pillow>=10.0.0

# Utilities
colorama
tqdm
psutil
